ref: Refactor get_client_config
The main goal is to take each variable that was determined by business
logic in the function body and encapsulate it in a property of the new
class. This makes it easier to tell at a glance what determines the
final value of each of those variables; previously, there were problems
with values such as `enabled_features` and `links` being modified
significantly below where they were first defined.

The side effect of calling `_delete_activeorg` is extracted to the
top-level `get_client_config` function. Everything that happens in
`_ClientConfig` is idempotent.