feat(backup): Use ordinals instead of pks
Prior to this change, model uniqueness was established using the
concatenation of model name ("sentry.somemodel") and the pk that
particular instance contains in the database. Unfortunately, this is not
sufficient for comparing two JSON dumps for equality, as the pks may
change depending on which underlying validation database is used.

This change moves us to a model where, at ingestion time, each model
being compared is assigned a per-model-kind ordinal, such that its
unique identifier is roughly "this is the 3rd lowest pk of model
`sentry.somemodel`".

To help this transition along, we simplify some of the equality checks,
reducing them to `UnorderedInput` and `UnequalCounts` (there are more
instance of a certain model kind in one side than the other).

Issue: getsentry/team-ospo#171